Renamed useMovies file to useMovies.ts (I Should've done that a long time ago).
